Move relevant routes from protocol to list
