+++ library_registry/library_registration_protocol/routes.py
@@ -1,9 +1,7 @@
 from flask import Blueprint, current_app
 
 from library_registry.decorators import (
-    compressible,
     has_library,
-    returns_json_or_response_or_problem_detail,
     returns_problem_detail,
     uses_location,
 )
@@ -45,26 +43,6 @@ def search_qa(_location):
 @returns_problem_detail
 def confirm_resource(resource_id, secret):
     return current_app.library_registry.validation_controller.confirm(resource_id, secret)
-
-@libr.route('/libraries')
-@compressible
-@uses_location
-@returns_problem_detail
-def libraries_opds(_location=None):
-    return current_app.library_registry.registry_controller.libraries_opds(location=_location)
-
-@libr.route('/libraries/qa')
-@compressible
-@uses_location
-@returns_problem_detail
-def libraries_qa(_location=None):
-    return current_app.library_registry.registry_controller.libraries_opds(location=_location, live=False)
-
-@libr.route('/library/<uuid>')
-@has_library
-@returns_json_or_response_or_problem_detail
-def library():
-    return current_app.library_registry.registry_controller.library()
 
 @libr.route('/library/<uuid>/eligibility')
 @has_library
